requirements.txt
pysubs2
google-cloud-texttospeech
regex
click
click-option-group
